Add portfolio evaluation export endpoint
